Switch db library to JayDeBeApi (#3)
The database to be used for testing is H2. To make the connection
portable between databases, this library is used. The connection
arguments are gathered from environment variables. There are .env
files used for each database. For MySQL, the .env file is used.
For H2 testing, the test/.env.test file is used.

Test can be run with the test/runtests.sh script. In that file,
evironment variables are set for the .env file for the test, and
also the Java classpath so H2 can find a required library for
adding MySQL functions to H2.