Revert 11 order user json xml (#12)
* Switch db library to JayDeBeApi (#3)

The database to be used for testing is H2. To make the connection
portable between databases, this library is used. The connection
arguments are gathered from environment variables. There are .env
files used for each database. For MySQL, the .env file is used.
For H2 testing, the test/.env.test file is used.

Test can be run with the test/runtests.sh script. In that file,
evironment variables are set for the .env file for the test, and
also the Java classpath so H2 can find a required library for
adding MySQL functions to H2.

* Update README

* Reactor user producer for testing and add tests (#4)

* Refactor to use `main` entrypoint function
* Add more user producer tests
* Add details to README

* Update README

* Add password hashing to user passwords (#6)

* Reactor to extract reusable code

Extract code for printing items and getting user confirmation.
This function should use used for all producers, in order to keep
the user experience uniform.

* Add Bcrypt dependency to requirements.txt

* Remove h2/lib from .gitignore

* Update schema and fix code accordingly (#7)

* Update models and fix save methods
* Update tests to match schema
* Update testdata to make dummy restaurant data

* Update schema (fields renamed) (#8)

* All table ids have been update to just be 'id'
* Some `binary(16)` have been made into `bigint(8) ai`

* Add json and xml support for user and order ingestion (#11)

* Refactor formatters to use a base abstract class
* Refactor producers to user base abstract class
* Add json support for ingestion
* Add xml support for ingestion
* Make ingestion and producers two different programs

* Added ingests and updated producers for new schema (#9)

* Added producers
* Added producers
* Fixed for h2 and got new database structure working
* Restructured files, updated for schema, added tests

* Revert "Add json and xml support for user and order ingestion (#11)"

This reverts commit e5a68e7c3d7a70152955e1824aadbae50d5d09b3.

Co-authored-by: RobertOlvisSmoothstack <89948267+RobertOlvisSmoothstack@users.noreply.github.com>